+++ src/client/request/graphql/generators/configuration.js
@@ -14,6 +14,8 @@ var CONFIGURATION_QUERY =
   "      threeDSecureEnabled " +
   "      threeDSecure { " +
   "        cardinalAuthenticationJWT " +
+  "        cardinalSongbirdUrl " +
+  "        cardinalSongbirdIdentityHash " +
   "      } " +
   "    } " +
   "    applePayWeb { " +