+++ src/three-d-secure/shared/constants.js
@@ -2,8 +2,4 @@
 
 module.exports = {
   LANDING_FRAME_NAME: "braintreethreedsecurelanding",
-  CARDINAL_SCRIPT_SOURCE: {
-    production: "https://songbird.cardinalcommerce.com/edge/v1/songbird.js",
-    sandbox: "https://songbirdstag.cardinalcommerce.com/edge/v1/songbird.js",
-  },
 };